feat: 진도율 기능 구현중

+++ src/main/java/com/sesac/backend/lectures/dto/request/ProgressRequest.java
@@ -8,6 +8,6 @@ import lombok.AllArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProgressRequest {
-    private Integer progressRate;
+    private Double progressRate;
     private Integer watchedSeconds;
 }
